Create 022Problem
A and B play games with 10 numbers. If win, the winner will get 3 points, and if lost, there is no point, and if tie, they will get 1 point for each other. 
A draw is when all 10 rounds are tied. When the score is the same, the last winner in the game will win.