Create Sort006(LRU)
Least Recently Used.
Cache Miss: If the CPU uses the new task, task 5, with the task to be done out of cache, the cache miss, all the tasks are pushed back, and task 5 is placed at the front of the cache.
Cache hit: If a cache has task 3 in it. When the CPU uses task 3, it becomes Cache Hit, and tasks 5 and 2 in front of task 3 are pushed back one space, and task 3 is in the front.